requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41
alembic==1.16.2
Mako==1.3.10
greenlet==3.2.3
typing-extensions==4.14.0
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
colorama==0.4.6
